move SLO-config interface to shared folder
